requirements.txt
-f https://download.pytorch.org/whl/cpu
torch
streamlit==1.42.1
matplotlib==3.10.0
ansi2html==1.9.2
pyvis==0.3.2
langchain==0.3.18
einops==0.8.1
ollama==0.4.7
python-docx==1.1.2
qdrant-client==1.13.2
crewai[tools]==0.102.0
pdfplumber
